Rebrand and restructure into 'consus-core'
