+++ Gruntfile.js
@@ -35,16 +35,11 @@ module.exports = function(grunt) {
             options: {
                 reporter: 'spec'
             },
-            unit: {
+            all: {
                 files: {
-                    src: ['.test/unit/**/*.js']
+                    src: ['.test/**/*.js']
                 }
             },
-            functional: {
-                files: {
-                    src: ['.test/functional/**/*.js']
-                }
-            }
         },
         clean: {
             coverage: ['coverage/', 'coverage.lcov', '.nyc_output/'],
