+++ README.md
@@ -1,71 +1,15 @@
-# consus-flux
+# consus-core
 
-[![Build Status](https://travis-ci.org/TheFourFifths/consus-flux.svg?branch=dev)](https://travis-ci.org/TheFourFifths/consus-flux)
-[![codecov](https://codecov.io/gh/TheFourFifths/consus-flux/branch/dev/graph/badge.svg)](https://codecov.io/gh/TheFourFifths/consus-flux)
-[![npm](https://img.shields.io/npm/v/consus-flux.svg)](https://www.npmjs.com/package/consus-flux)
-[![devDependency Status](https://david-dm.org/TheFourFifths/consus-flux/dev-status.svg)](https://david-dm.org/TheFourFifths/consus-flux?type=dev)
+[![Build Status](https://travis-ci.org/TheFourFifths/consus-core.svg?branch=dev)](https://travis-ci.org/TheFourFifths/consus-core)
+[![codecov](https://codecov.io/gh/TheFourFifths/consus-core/branch/dev/graph/badge.svg)](https://codecov.io/gh/TheFourFifths/consus-core)
+[![npm](https://img.shields.io/npm/v/consus-core.svg)](https://www.npmjs.com/package/consus-core)
+[![devDependency Status](https://david-dm.org/TheFourFifths/consus-core/dev-status.svg)](https://david-dm.org/TheFourFifths/consus-core?type=dev)
 
-Flux modules for the Consus project
+Common modules for the Consus project
 
 ## Installing
 
-`npm install consus-flux --save`
-
-## Using the Dispatcher
-
-```javascript
-import { Dispatcher } from 'consus-flux';
-
-Dispatcher.handleAction('INCREMENT', {
-    amount: 5
-});
-```
-
-## Using the Store
-
-```javascript
-import { Store } from 'consus-flux';
-
-let count = 0;
-
-class CounterStore extends Store {
-
-    getCount() {
-        return count;
-    }
-
-}
-
-const store = new CounterStore();
-
-store.registerHandler('INCREMENT', data => {
-    count += data.amount;
-    store.emitChange();
-});
-
-store.registerHandler('DECREMENT', data => {
-    count -= data.amount;
-    store.emitChange();
-});
-
-export default store;
-```
-
-## Listening to a Store
-
-```javascript
-import CounterStore from './counter-store';
-
-function handleChange() {
-    console.log('The count is now: ' + CounterStore.getCount());
-}
-
-CounterStore.addChangeListener(handleChange);
-
-setTimeout(() {
-    CounterStore.removeChangeListener(handleChange);
-}, 10000);
-```
+`npm install consus-core --save`
 
 ## Developing
 
@@ -73,9 +17,9 @@ setTimeout(() {
 
 ```bash
 # Clone the repository
-git clone git@github.com:TheFourFifths/consus-flux.git
+git clone git@github.com:TheFourFifths/consus-core.git
 # Enter the project directory
-cd consus-flux
+cd consus-core
 # Install dependencies
 npm install
 # Build the project
@@ -93,7 +37,7 @@ npm test
 
 ### Project File Structure
 
+* `docs`: Project documentation goes here
 * `src`: The project's source code
 * `test`: The project's tests
-    * `lib`: Miscellaneous library modules
     * `unit`: Unit tests
