+++ index.js
@@ -1,0 +1,3 @@
+module.exports = {
+    flux: require('././flux');
+};
