+++ package.json
@@ -1,8 +1,8 @@
 {
-  "name": "consus-flux",
-  "description": "Flux modules for the Consus project",
-  "version": "0.3.0",
-  "author": "The Four Fiths",
+  "name": "consus-core",
+  "description": "Common modules for the Consus project",
+  "version": "0.4.0",
+  "author": "The Four Fifths",
   "contributors": [
     "Jordan Longato",
     "Matthew Mahnke",
@@ -11,7 +11,7 @@
     "Trevin Hofmann <trevinhofmann@gmail.com>"
   ],
   "license": "Unlicense",
-  "main": ".dist/index.js",
+  "main": "index.js",
   "scripts": {
     "build": "grunt build",
     "coverage": "cross-env NODE_ENV=test BABEL_ENV=test nyc npm test && nyc report --reporter=text-lcov > coverage.lcov",
@@ -21,14 +21,14 @@
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/TheFourFifths/consus-flux"
+    "url": "https://github.com/TheFourFifths/consus-core"
   },
   "bugs": {
-    "url": "https://github.com/TheFourFifths/consus-flux/issues"
+    "url": "https://github.com/TheFourFifths/consus-core/issues"
   },
   "keywords": [
     "consus",
-    "flux",
+    "core",
     "the four fifths"
   ],
   "devDependencies": {
