+++ src/flux/index.js
@@ -1,0 +1,7 @@
+import Dispatcher from './dispatcher';
+import Store from './store';
+
+module.exports = {
+    Dispatcher,
+    Store
+};
