--- src/index.js
@@ -1,7 +1,0 @@
-import Dispatcher from './lib/dispatcher';
-import Store from './lib/store';
-
-module.exports = {
-    Dispatcher,
-    Store
-};
